+++ tests/qiskit_convert_test.py
@@ -1,4 +1,4 @@
-# Copyright 2019-2024 Quantinuum
+# Copyright Quantinuum
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -32,11 +32,11 @@ from qiskit.circuit.equivalence_library import (  # type: ignore
 from qiskit.circuit.library import (
     MCMTGate,
     PauliEvolutionGate,
-    RealAmplitudes,
     RYGate,
-    TwoLocal,
     UnitaryGate,
     XXPlusYYGate,
+    n_local,
+    real_amplitudes,
 )
 from qiskit.circuit.parameterexpression import ParameterExpression  # type: ignore
 from qiskit.quantum_info import Operator, SparsePauliOp, Statevector  # type: ignore
@@ -1124,22 +1124,17 @@ def test_failed_conversion_error() -> None:
 
 
 # https://github.com/CQCL/pytket-qiskit/issues/200
-def test_RealAmplitudes_numeric_params() -> None:
+def test_real_amplitudes_numeric_params() -> None:
     qc = QuantumCircuit(3)
     params = [np.pi / 2] * 9
-    real_amps1 = RealAmplitudes(3, reps=2)
+    real_amps1 = real_amplitudes(3, reps=2)
     real_amps2 = real_amps1.assign_parameters(params)
     qc.compose(real_amps2, qubits=[0, 1, 2], inplace=True)
     # Unitary operator of the qiskit circuit. Order reversed from little -> big endian.
     # The reversal means we can check it for equivalence with a tket unitary
     qiskit_unitary = Operator(qc.reverse_bits()).data
     converted_tkc = qiskit_to_tk(qc)
-    assert converted_tkc.n_gates == 1
-    assert converted_tkc.n_gates_of_type(OpType.CircBox) == 1
-    circbox_op = converted_tkc.get_commands()[0].op
-    assert isinstance(circbox_op, CircBox)
-    assert circbox_op.get_circuit().name == "RealAmplitudes"
-    DecomposeBoxes().apply(converted_tkc)
+    assert converted_tkc.n_gates == 13
     assert converted_tkc.n_gates_of_type(OpType.CX) == 4
     assert converted_tkc.n_gates_of_type(OpType.Ry) == 9
     unitary1 = converted_tkc.get_unitary()
@@ -1152,7 +1147,7 @@ def test_RealAmplitudes_numeric_params() -> None:
 
 # https://github.com/CQCL/pytket-qiskit/issues/256
 def test_symbolic_param_conv() -> None:
-    qc = TwoLocal(1, "ry", "cz", reps=1, entanglement="linear")
+    qc = n_local(2, "ry", "cz", reps=1, entanglement="linear")
     qc_transpiled = transpile(
         qc, basis_gates=["sx", "rz", "cx", "x"], optimization_level=3
     )
@@ -1191,6 +1186,113 @@ def test_nonregister_bits() -> None:
     c.rename_units({Bit(0): Bit(1)})
     with pytest.raises(NotImplementedError):
         tk_to_qiskit(c)
+
+
+# https://github.com/CQCL/pytket-qiskit/issues/415
+def test_ifelseop_two_branches() -> None:
+    qreg = QuantumRegister(1, "r")
+    creg = ClassicalRegister(1, "s")
+    circuit = QuantumCircuit(qreg, creg)
+
+    circuit.h(qreg[0])
+    circuit.measure(qreg[0], creg[0])
+
+    with circuit.if_test((creg[0], 1)) as else_:
+        circuit.h(qreg[0])
+    with else_:
+        circuit.x(qreg[0])
+    circuit.measure(qreg[0], creg[0])
+
+    tkc = qiskit_to_tk(circuit)
+    tkc.name = "test_circ"
+
+    # Manually build the expected pytket Circuit.
+    # Validate against tkc.
+    expected_circ = Circuit(name="test_circ")
+    r_reg = expected_circ.add_q_register("r", 1)
+    s_reg = expected_circ.add_c_register("s", 1)
+    expected_circ.H(r_reg[0])
+    expected_circ.Measure(r_reg[0], s_reg[0])
+
+    h_circ = Circuit()
+    h_reg = h_circ.add_q_register("r", 1)
+    h_circ.name = "If"
+    h_circ.H(h_reg[0])
+
+    x_circ = Circuit()
+    x_reg = x_circ.add_q_register("r", 1)
+    x_circ.name = "Else"
+    x_circ.X(x_reg[0])
+
+    expected_circ.add_circbox(
+        CircBox(h_circ), [r_reg[0]], condition_bits=[s_reg[0]], condition_value=1
+    )
+    expected_circ.add_circbox(
+        CircBox(x_circ), [r_reg[0]], condition_bits=[s_reg[0]], condition_value=0
+    )
+
+    expected_circ.Measure(r_reg[0], s_reg[0])
+
+    assert expected_circ == tkc
+
+
+# https://github.com/CQCL/pytket-qiskit/issues/415
+def test_ifelseop_one_branch() -> None:
+    qubits = QuantumRegister(1, "q1")
+    clbits = ClassicalRegister(1, "c1")
+    circuit = QuantumCircuit(qubits, clbits)
+    (q0,) = qubits
+    (c0,) = clbits
+
+    circuit.h(q0)
+    circuit.measure(q0, c0)
+    with circuit.if_test((c0, 1)):
+        circuit.x(q0)
+    circuit.measure(q0, c0)
+
+    tket_circ_if_else = qiskit_to_tk(circuit)
+    tket_circ_if_else.name = "test_circ"
+
+    # Manually build the expected pytket Circuit.
+    # Validate against tket_circ_if_else.
+    expected_circ = Circuit()
+    expected_circ.name = "test_circ"
+    q1_tk = expected_circ.add_q_register("q1", 1)
+    c1_tk = expected_circ.add_c_register("c1", 1)
+    expected_circ.H(q1_tk[0])
+    expected_circ.Measure(q1_tk[0], c1_tk[0])
+    x_circ = Circuit()
+    x_circ.name = "If"
+    xq1 = x_circ.add_q_register("q1", 1)
+    x_circ.X(xq1[0])
+    expected_circ.add_circbox(
+        CircBox(x_circ), [q1_tk[0]], condition_bits=[c1_tk[0]], condition_value=1
+    )
+
+    expected_circ.Measure(q1_tk[0], c1_tk[0])
+
+    assert tket_circ_if_else == expected_circ
+
+
+def test_ifelseop_multi_bit_cond() -> None:
+    qreg = QuantumRegister(2, "q")
+    creg = ClassicalRegister(2, "c")
+    circuit = QuantumCircuit(creg, qreg)
+    (q0, q1) = qreg
+    (c0, c1) = creg
+
+    circuit.h(q0)
+    circuit.h(q1)
+    circuit.measure(q0, c0)
+    circuit.measure(q1, c1)
+    with circuit.if_test((creg, 2)):
+        circuit.x(q0)
+        circuit.x(q1)
+    circuit.measure(q0, c0)
+    circuit.measure(q1, c1)
+    # This currently gives an error as register exp not supported.
+    with pytest.raises(NotImplementedError):
+        qiskit_to_tk(circuit)
 
 
 def test_range_preds_with_conditionals() -> None: