+++ docs/changelog.md
@@ -7,6 +7,7 @@
 ## 0.63.0 (January 2025) - UNRELEASED
 
 - Support conversion of qiskit circuits containing `IfElseOp` in the {py:func}`qiskit_to_tk` converter.
+- Reject circuits containing nested conditionals when converting to qiskit.
 
 ## 0.62.0 (December 2024)