+++ docs/changelog.md
@@ -4,9 +4,12 @@
 
 # Changelog
 
-## Unreleased
+## 0.63.0 (January 2025)
 
+- Support conversion of qiskit circuits containing `IfElseOp` in the {py:func}`qiskit_to_tk` converter. Note: Only conditions on a single bit are supported currently. Handling of more general conditions will be added in future.
 - Reject circuits containing nested conditionals when converting to qiskit.
+- Update pytket version requirement to 1.39.0. 
+- Add support for fractional gates on backends that support them.
 
 ## 0.62.0 (December 2024)